[sse] Refactor. Remove process exit on 403 sse errors
