+++ src/modules/utils.ts
@@ -10,7 +10,7 @@ import * as env from '../env'
 import { BuildFailError } from '../errors'
 import log from '../logger'
 import { getAppRoot } from '../manifest'
-import { logAll, onEvent } from '../sse'
+import { logAll, onEvent } from '../lib/sse'
 import { createTable } from '../table'
 import userAgent from '../user-agent'
 import { promptConfirm } from './prompts'
